more MultiplierToTableMigrator tests
