+++ java_console/ui/src/test/java/com/rusefi/maintenance/migration/multiplier_to_table_migration/test_data/cranking_fuel_multiplier/prev_calibrations.ini
@@ -1,0 +1,14 @@
+[MegaTune]
+signature       = "rusEFI afr_msq_import_migration_test_afr_to_lambda_previous"
+
+[Constants]
+pageReadCommand = "R%2o%2c"
+crc32CheckCommand	= "k%2i%2o%2c"
+pageSize        = 46496
+page            = 1
+cranking_baseFuel = scalar, F32, 180, "mg", 1, 0, 0, 500, 1
+crankingCycleBins = array, F32, 13888, [8], "counter", 1, 0, -80, 170, 0
+crankingCycleFuelCltBins = array, S16, 13920, [4], "C", 1, 0, -100, 250, 2
+crankingCycleFuelCoef = array, F32, 13928, [8x4], "mult", 1, 0, 0, 5, 2
+[OutputChannels]
+ochBlockSize    = 2008
