+++ java_console/ui/src/test/java/com/rusefi/maintenance/migration/multiplier_to_table_migration/test_data/cranking_fuel_multiplier/updated_calibrations.ini
@@ -1,0 +1,14 @@
+[MegaTune]
+signature       = "rusEFI afr_msq_import_migration_test_afr_to_lambda_updated"
+
+[Constants]
+crc32CheckCommand	= "k%2i%2o%2c"
+pageReadCommand = "R%2o%2c"
+pageSize        = 47572
+page            = 1
+crankingCycleBins = array, F32, 14008, [8], "counter", 1, 0, -80, 170, 0
+crankingCycleFuelCltBins = array, S16, 14040, [4], "C", 1, 0, -100, 250, 2
+crankingCycleBaseFuel = array, F32, 14048, [8x4], "mg", 1, 0, 0, 500, 1
+
+[OutputChannels]
+ochBlockSize    = 2064
